+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Added a `proto migrate` command for easily applying changes between breaking releases.
 - Added support for minisign checksum files. Can now verify `.minisig` signatures for downloaded tools.
 - Updated `proto use` to install tools in parallel.
+- Updated `proto plugins` and `proto tools` to load plugins in parallel.
 - TOML API
   - Added an `install.checksum_public_key` setting.
 - WASM API